Métodos de Deletar, Cadastrar e BuscarPorId implementados

+++ BackEnd/API2RP_wepAPI/Domains/Tiposusuario.cs
@@ -10,7 +10,7 @@ namespace API2RP_wepAPI.Domains
             Usuarios = new HashSet<Usuario>();
         }
 
-        public short IdTipoUsuario { get; set; }
+        public int IdTipoUsuario { get; set; }
         public string Titulo { get; set; } = null!;
 
         public virtual ICollection<Usuario> Usuarios { get; set; }
