+++ BackEnd/API2RP_wepAPI/Domains/Usuario.cs
@@ -5,8 +5,8 @@ namespace API2RP_wepAPI.Domains
 {
     public partial class Usuario
     {
-        public short IdUsuario { get; set; }
-        public short IdTipoUsuario { get; set; }
+        public int IdUsuario { get; set; }
+        public int IdTipoUsuario { get; set; }
         public string Nome { get; set; } = null!;
         public string Email { get; set; } = null!;
         public string Senha { get; set; } = null!;
